nova versão para IE_KNIN_C

+++ Cinterpolation_permutation/interpolation.h
@@ -5,7 +5,6 @@
 #ifndef CINTERPOLATION_PERMUTATION_INTERPOLATION_H
 #define CINTERPOLATION_PERMUTATION_INTERPOLATION_H
 
-#include "./../math/Point.h"
-
-double * interpolation_mount(struct Point * Pares);
+#include "point.h"
+long double * interpolation_mount(struct Point * Pares);
 #endif //CINTERPOLATION_PERMUTATION_INTERPOLATION_H
